modularize: Config and run to lib.rs
